pyproject.toml
[project]
name = "open-alpha-evolve"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "arize-phoenix>=11.1.1",
    "arize-phoenix-otel>=0.12.1",
    "openinference-instrumentation-openai>=0.1.30",
]
